+++ catscripts/Entertainment/Movies/Jurassic_Park_movies/JurassicPark.sh
@@ -2,15 +2,23 @@
 
 export KEYWORDS_MOVIES_JURASSIC_PARK="Jur(|r)as(|s)ic(| )Park"
 
-egrep -i "$KEYWORDS_MOVIES_JURASSIC_PARK" newpages.txt >> JurassicPark.txt
-
-JURASSICPARK=$(stat --print=%s JurassicPark.txt)
-
-if [ $JURASSICPARK -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="JurassicPark.txt"
-  export CATNAME="Jurassic Park"
-  $CATEGORIZE
-fi
 
-rm JurassicPark.txt
+  debug_start "Jurassic Park"
+
+  JURASSICPARK=$(egrep -i "$KEYWORDS_MOVIES_JURASSIC_PARK" newpages.txt)
+
+  if [ "$JURASSICPARK" != "" ];
+  then
+    printf "%s" "$JURASSICPARK" > JurassicPark.txt
+    export CATFILE="JurassicPark.txt"
+    export CATNAME="Jurassic Park"
+    $CATEGORIZE
+    rm JurassicPark.txt
+    unset JURASSICPARK
+  fi
+
+  debug_end "Jurassic Park"
+
+fi
