+++ catscripts/Entertainment/Movies/The_Hunger_Games/HungerGames.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i 'Hunger Game|Peeta|Everdeen|Katniss|Haymitch|Nightlock' newpages.txt >> TheHungerGames.txt
+KEYWORDS_HUNGERGAMES="Hunger(| )Game|Peeta|Everdeen|Katniss|Haymitch|Nightlock"
 
-HUNGER=$(stat --print=%s TheHungerGames.txt)
-
-if [ $HUNGER -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="TheHungerGames.txt"
-  export CATNAME="The Hunger Games"
-  $CATEGORIZE
-fi
 
-rm TheHungerGames.txt
+  debug_start "The Hunger Games"
+
+  HUNGER=$(egrep -i "$KEYWORDS_HUNGERGAMES" newpages.txt)
+
+  if [ "$HUNGER" != "" ];
+  then
+    printf "%s" "$HUNGER" > TheHungerGames.txt
+    export CATFILE="TheHungerGames.txt"
+    export CATNAME="The Hunger Games"
+    $CATEGORIZE
+    rm TheHungerGames.txt
+    unset HUNGER
+  fi
+
+  debug_end "The Hunger Games"
+
+fi
