+++ catscripts/Entertainment/Video_games/MMORPGs/Guild_Wars/GuildWars.sh
@@ -2,15 +2,23 @@
 
 export KEYWORDS_MMORPG_GUILDWARS="Guild(| )War|PVX"
 
-egrep -i "$KEYWORDS_MMORPG_GUILDWARS" newpages.txt >> GuildWars.txt
-
-GUILDWARS=$(stat --print=%s GuildWars.txt)
-
-if [ $GUILDWARS -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="GuildWars.txt"
-  export CATNAME="Guild Wars"
-  $CATEGORIZE
-fi
+  
+  debug_start "Guild Wars"
 
-rm GuildWars.txt
+  GUILDWARS=$(egrep -i "$KEYWORDS_MMORPG_GUILDWARS" newpages.txt)
+
+  if [ "$GUILDWARS" != "" ];
+  then
+    printf "%s" "$GUILDWARS" > GuildWars.txt
+    export CATFILE="GuildWars.txt"
+    export CATNAME="Guild Wars"
+    $CATEGORIZE
+    rm GuildWars.txt
+    unset GUILDWARS
+  fi
+
+  debug_end "Guild Wars"
+
+fi
