+++ catscripts/Entertainment/Video_games/MMORPGs/Star_Wars_The_Old_Republic/SWTOR.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 
 export KEYWORDS_MMORPG_SWTOR="The(| )Old(| )Republic|SW(:|)TOR"
+KEYWORDS_MMORPG_SWTOR_EXCLUDE="Knights(| )of(| )the(| )old(| )Republic"
 
-egrep -i "$KEYWORDS_MMORPG_SWTOR" newpages.txt | egrep -iv 'Knights(| )of(| )the(| )old(| )Republic' >> SWTOR.txt
-
-SWTOR=$(stat --print=%s SWTOR.txt)
-
-if [ $SWTOR -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="SWTOR.txt"
-  export CATNAME="Star Wars: The Old Republic"
-  $CATEGORIZE
-fi
 
-rm SWTOR.txt
+  debug_start "Star Wars: The Old Republic"
+
+  SWTOR=$(egrep -i "$KEYWORDS_MMORPG_SWTOR" newpages.txt | egrep -iv "$KEYWORDS_MMORPG_SWTOR_EXCLUDE")
+
+  if [ "$SWTOR" != "" ];
+  then
+    printf "%s" "$SWTOR" > SWTOR.txt
+    export CATFILE="SWTOR.txt"
+    export CATNAME="Star Wars: The Old Republic"
+    $CATEGORIZE
+    rm SWTOR.txt
+    unset SWTOR
+  fi
+
+  debug_end "Star Wars: The Old Republic"
+
+fi
