+++ catscripts/Entertainment/Video_games/MMORPGs/World_of_Warcraft/WorldOfWarcraft.sh
@@ -1,17 +1,25 @@
 #!/bin/bash
 
 export KEYWORDS_MMORPG_WOW="World(| )of(| )Warcraft|w\.o\.w"
+KEYWORDS_MMORPG_WOW_CASESENSITIVE="WoW"
 
-egrep -i "$KEYWORDS_MMORPG_WOW" newpages.txt >> WorldOfWarcraft.txt
-egrep 'WoW' newpages.txt >> WorldOfWarcraft.txt
-
-WOW=$(stat --print=%s WorldOfWarcraft.txt)
-
-if [ $WOW -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="WorldOfWarcraft.txt"
-  export CATNAME="World of Warcraft"
-  $CATEGORIZE
-fi
 
-rm WorldOfWarcraft.txt
+  debug_start "World of Warcraft"
+
+  WOW=$(egrep -i "$KEYWORDS_MMORPG_WOW" newpages.txt && egrep "$KEYWORDS_MMORPG_WOW_CASESENSITIVE" newpages.txt)
+
+  if [ "$WOW" != "" ];
+  then
+    printf "%s" "$WOW" > WorldOfWarcraft.txt
+    export CATFILE="WorldOfWarcraft.txt"
+    export CATNAME="World of Warcraft"
+    $CATEGORIZE
+    rm WorldOfWarcraft.txt
+    unset WOW
+  fi
+
+  debug_end "World of Warcraft"
+
+fi
