+++ catscripts/Entertainment/Video_games/Role-playing_video_games/Star_Wars_Knights_of_the_Old_Republic/KOTOR.sh
@@ -1,14 +1,25 @@
 #!/bin/bash
 
-egrep -i 'Knights of the Old Republic|KOTOR' newpages.txt | egrep -iv 'The(| )Old(| )Republic(| )(2|II)|KOTOR(| )(2|II)' >> KOTOR.txt
+KEYWORDS_KOTOR="Knights(| )of(| )the(| )Old(| )Republic|KOTOR"
+KEYWORDS_KOTOR_EXCLUDE="The(| )Old(| )Republic(| )(2|II)|KOTOR(| )(2|II)"
 
-KOTOR=$(stat --print=%s KOTOR.txt)
-
-if [ $KOTOR -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="KOTOR.txt"
-  export CATNAME="Star Wars: Knights of the Old Republic"
-  $CATEGORIZE
-fi
 
-rm KOTOR.txt
+  debug_start "Star Wars: Knights of the Old Republic"
+
+  KOTOR=$(egrep -i "$KEYWORDS_KOTOR" newpages.txt | egrep -iv "$KEYWORDS_KOTOR_EXCLUDE")
+
+  if [ "$KOTOR" != "" ];
+  then
+    printf "%s" "$KOTOR" > KOTOR.txt
+    export CATFILE="KOTOR.txt"
+    export CATNAME="Star Wars: Knights of the Old Republic"
+    $CATEGORIZE
+    rm KOTOR.txt
+    unset KOTOR
+  fi
+
+  debug_end "Star Wars: Knights of the Old Republic"
+
+fi
